requirements.txt
discord.py==2.3.2
pylint==3.0.2
radon==6.0.1
bandit==1.7.5
